PWGHF: Add a separate data model for correlation tables (#3642)
